Update branding, documentation, and network configuration
- Updated all branding materials: replaced "Maddison Lane" with "Neil Adamson"
- Removed WCAG compliance references from brand guidelines and color palette
- Hardcoded genesis block values (nonce: 1827011, merkle root, randomx key)
- Fixed cross-platform build: GCC-specific compiler flags now conditional
- Updated I2P destination address to deployed server address
- Corrected network port numbers in README (mainnet P2P: 2211, faucet: 2215)
- Removed config/ and tools/ directories (no longer needed)

+++ src/pool/pool.cpp
@@ -481,7 +481,7 @@ uint64_t CalculateShareDifficulty(const uint256& hash) {
     }
 
     // Minimum difficulty of 1
-    return std::max(difficulty, 1ULL);
+    return std::max(difficulty, static_cast<uint64_t>(1));
 }
 
 uint256 GenerateJobID() {
